Change to SPA for the game page since we're communicating on both via the websocket. Stub out the basic role screen. Break out the queue screen and logic for it. Update the main page to remove the queue from the url
